+++ src/graphics/renderables/styled_info.ts
